Pattern overlap checker via unification
This will allow us to properly handle type args for reachability of type
patterns.

Also moves the unification logic core shared with coherence to its own
module
